Move utility functions to common module. Remove test programs.
RISC-V test programs exist in the tapasco-riscv repo already.
Remove some print statements.
