--- riscv-testprograms/.gitignore
@@ -1,1 +1,0 @@
-/*/elf
