--- riscv-testprograms/minimal/Makefile
@@ -1,3 +1,0 @@
-ROOTDIR ?= $(shell pwd)/..
-
-include $(ROOTDIR)/base.mk
